+++ src/body.c
@@ -1,6 +1,6 @@
 #include "body.h"
 
-void phys_init_body(phys_body* body, phys_vector2 pos, phys_vector2 vel, phys_vector2 accel, float mass) {
+void phys_init_body(phys_body* body, phys_vector2 pos, phys_vector2 vel, phys_vector2 accel, double mass) {
 	body->pos = pos;
 	body->vel = vel;
 	body->accel = accel;
